requirements.txt
django
psycopg2
django-fsm
python-dotenv